Created copy of haproxy-release
The haproxy-release (part of the cf-release) has been
copied to this repository, so we can extend its functionality.
See the README for why this was not forked. We have used the
same .gitignore as found in the cf-release.

We had to merge the common package into haproxy:

This release was composed of two packages (common and haproxy)
and one job. The common package is from the cf-release, which will
be installed on the same VM (router_z*). This causes a name conflict
as bosh cannot co-locate two packages of the same name on a VM.
The common package just includes a single bash script, so this
file has been placed under the haproxy package and the common
package has been removed.