Changed path of haproxy properties in spec
We want to keep all the manifest properties of the haproxy job inside
of the `haproxy` hash, so we won't refer to the `router` variables
anymore.

Despite this, we will still read the variable `cc.allow_app_ssh_access`
to keep configuring the tenant ssh feature.

Additionally, as we will use haproxy colocated in the same VM than the go
router, there is no need to distinguish z1 and z2 vms.