Allow define raw frontend configurations
Allow define raw frontend configuration for HAProxy, so the user of
this release can add HAProxy config directly in the manifest to
modify the HTTP request and response.

We change the two templates, haproxy.conf.erb and haproxy.ctmpl.erb.
The second one is used if the installation is configured to support
ssh access from the tenants and consul is installed in the system,
so consul will render the haproxy dynamically with the cell
backends.