Add new frontend for HTTP ProxyProtocol traffic
The new optional frontend can be enabled with haproxy.enable_http_proxy.

It will setup a frontend in port 81 which can understand
ProxyProtocol[1] without SSL. The frontend will have the same
config than the rest of frontends.

It will automatically setup the X-Forwarded-Proto if the client
connected to 443 in the previous LB.