Add monitoring frontend with custom User-Agent
In order to allow AWS ELBs to monitor gorouter and detect when it is
draining connections, we add an additional frontend on an alternate port
which inserts a User Agent [1] header on the request to backend.
Gorouter identifies this header, responding 503 if it is draining connections.
This way ELB will detect the node as unhealthy and remove it from the pool of
backends, so it will stop sending new connections.

[1] https://github.com/cloudfoundry/gorouter#instrumentation