Apex domain HSTS configuration
The HSTS configuration was previously configured in property
`ha_proxy.additional_frontend_config`. We now need a more customised
configuration because we want to handle the HSTS configuration for the
apex domain as well, which is slightly different than the rest.
The HAProxy configuration is now less generic but it was never really
generic anyway.

HSTS can now be enabled across all requests by setting property
`ha_proxy.hsts.enable` to true. The behaviour is the same as the one we
had when we were setting HSTS in additional_frontend_config from
paas-cf.

The preload configuration can now be enabled by setting
`ha_proxy.hsts.apex_domain` to the apex domain name. The requests to
apex use a different HAProxy backend to set a different headers, but
requests are still proxied to gorouter. Gorouter may respond with a 404
as there is probably no applications deployed to CF with that domain,
but the HSTS headers will still be set and this is enough for HSTS
preload.