Capture the Host header for logging
Currently we capture the access log of the HAProxy in front of our gorouters.

That is valuable information, but in order to identify for which
application each request is, we need to log the Host header.

To do so, we add the instruction `capture request header` for each
required frontend. The current log format `httplog` will print the list
of captured headers[1], so it is not necessary to change the format.

[1] http://blog.haproxy.com/2012/10/29/haproxy-log-customization/