Redirect all traffic to https protocol
We suffer from the lack of redirection in place, causing our users to be
confused when it comes to accessing their applications. We have tried
previously to set up the HSTS which has somewhat helped, however is very
limited to new versions of specific browsers.

With the following change, we're hoping to solve that issue, without the
use of external factors but handle the redirection on our servers.