Remove consul integration for haproxy config.
This used to use consul to populate the haproxy config from a template
with details of the SSH proxy servers. The ssh endpoints were removed in
e99e3f5, meaning that there are no longer any template operators in the
templace, and it's merely copied over the top of the .conf file.

We are therefore removing this feature because it only adds confusion
and duplication. It also has an implicit dependency on the consul
package.